requirements.txt
# Requirements for the project.

apache-airflow==2.7.1
requests
pandas
psycopg2-binary
sqlalchemy
urllib3
datetime
sqlalchemy-redshift
redshift_connector
apache-airflow-providers-postgres

